+++ src/CurlTransport.php
@@ -141,10 +141,9 @@ class CurlTransport extends Transport
         }
 
         $content = $request->getContent();
-        if ($content === null) {
-            if ($method === 'HEAD') {
-                $curlOptions[CURLOPT_NOBODY] = true;
-            }
+
+        if ($method === 'HEAD') {
+            $curlOptions[CURLOPT_NOBODY] = true;
         } else {
             $curlOptions[CURLOPT_POSTFIELDS] = $content;
         }