+++ tests/CurlTransportTest.php
@@ -2,6 +2,7 @@
 
 namespace yiiunit\extensions\httpclient;
 
+use yii\httpclient\Client;
 use yii\httpclient\CurlTransport;
 
 /**
@@ -61,5 +62,57 @@ class CurlTransportTest extends TransportTestCase
             CURLOPT_SSLCERTPASSWD => $options['sslPassphrase'],
         ];
         $this->assertEquals($expectedContextOptions, $contextOptions);
+    }
+
+    public function testPreparePostRequestWithEmptyBody()
+    {
+        $client = new Client([
+            'transport' => 'yii\httpclient\CurlTransport',
+        ]);
+        $request = $client->createRequest();
+        $request->setMethod('POST');
+        $request->setUrl('http://app.test/full/url');
+
+        $transport = $this->createClient()->getTransport();
+        $curlOptions = $this->invoke($transport, 'prepare', [$request]);
+
+        $expectedCurlOptions = [
+            CURLOPT_POST => true,
+            CURLOPT_POSTFIELDS => null,
+            CURLOPT_RETURNTRANSFER => true,
+            CURLOPT_URL => 'http://app.test/full/url',
+            CURLOPT_HTTPHEADER => [
+                0 => 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8',
+                1 => 'Content-Length: 0',
+            ],
+        ];
+
+        $this->assertEquals($expectedCurlOptions, $curlOptions);
+    }
+
+    public function testPrepareHeadRequestShouldNotHaveBody()
+    {
+        $client = new Client([
+            'transport' => 'yii\httpclient\CurlTransport',
+        ]);
+        $request = $client->createRequest();
+        $request->setMethod('HEAD');
+        $request->setUrl('http://app.test/full/url');
+
+        $transport = $this->createClient()->getTransport();
+        $curlOptions = $this->invoke($transport, 'prepare', [$request]);
+
+        $expectedCurlOptions = [
+            CURLOPT_RETURNTRANSFER => true,
+            CURLOPT_URL => 'http://app.test/full/url',
+            CURLOPT_HTTPHEADER => [
+                0 => 'Content-Type: application/x-www-form-urlencoded; charset=UTF-8',
+                1 => 'Content-Length: 0',
+            ],
+            CURLOPT_CUSTOMREQUEST => 'HEAD',
+            CURLOPT_NOBODY => true,
+        ];
+
+        $this->assertEquals($expectedCurlOptions, $curlOptions);
     }
 }