+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 HTTP client extension Change Log
 2.0.9 under development
 -----------------------
 
+- Bug #171: Add "Content-Length: 0" header when sending request with empty body (alexkart)
 - Bug #149: Fix type error in `StreamTransport` when `$http_response_header = null` (alexkart)